Pull

+++ src/Item.java
@@ -1,15 +1,14 @@
+// This abstract class Item is a base for subclasses Readable and Audio. 
 
 public abstract class Item{
-		
 	
-		
-		protected int sNo;
-		protected String title;
-		protected int price;
-		protected int quantity;
-		protected String itemtype;
-//		public abstract String[] getInfo(String filename, int lineNumber);
-		public abstract double getPrice();
+	// these are properties common to all Item Objects
+		protected int sNo; // sNo is the serial number of the item
+		protected String title; // the title of the item
+		protected double price; // the base price of the item
+		protected int quantity; // the quantity of the item
+		protected String itemtype; // the type of the item (can be CD, MP3, Book, eBook)
+		public abstract double getPrice(); // gets the price of the item
 		
 }
